Move code to $lib
