+++ packages/create-web/template/web/src/lib/charts/ApiBarChart.svelte
