+++ packages/create-web/template/web/src/lib/charts/BarChart.svelte
