+++ packages/create-web/template/web/src/lib/charts/BubbleChart.svelte
