+++ packages/create-web/template/web/src/lib/python_worker.ts
