+++ packages/create-web/template/web/src/lib/rust_worker.ts
